Debounce drag and drop Improve perceived performance by moving the drag/drop placeholder to the cursor, rather than one line at a time.
This two approaches kindof need one another. Once debounced, there aren't enough mouse-move events to get the placeholder to where it needs to be, especially on a fast cursor move. BUT this approach minimises event thrashing and tree recomputations.

I about to test this on a decently sized tree (about 1000 nodes), so stay tuned for further commits.