events now call with much more information. expand / collapse node now calls onToggleCollapse, with the tree, and the node that's changed. onChange events for dragging now fire with the following args: tree, node, parent, prev, next
the whole tree gets a 'currently-dragging' class when you are mid-drag. the idea is that if you have quite complex tree nodes, you can use CSS to hide some complexity when the user is dragging, and to make drag and drop a bit nicer.

I've also updated the drag handlers so that the item will move directly to the cursor. This means that if a system is firing mouse events infrequently, the placeholder doesn't get 'stuck' halfway down a long tree on a rapid mousemove.